third_party/blink/web_tests/platform/generic/svg/dom/style-reflect-expected.txt

This test checks that the type, media and title properties on an SVGStyleElement reflect the corresponding attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.type is "text/css"
PASS style.media is "all"
PASS style.title is ""
PASS style.type is "text/x-something-else"
PASS style.media is "print"
PASS style.title is "My Style Sheet"
PASS successfullyParsed is true

TEST COMPLETE
